Debugged

+++ Training_LFADS.py
@@ -30,7 +30,7 @@ def train_lfads(
         kl_weight = min(kl_end, kl_start + (kl_end - kl_start) * epoch / kl_anneal_epochs)
 
         for xb in train_loader:
-            xb = xb.to(device)
+            xb = xb.to(device).float()
             rates, kl_ic, kl_ctrl, _ = model(xb)
             loss, rec = lfads_loss(rates, xb, kl_ic, kl_ctrl, kl_weight)
 
@@ -47,7 +47,7 @@ def train_lfads(
         with torch.no_grad():
             val_loss = 0
             for xb in val_loader:
-                xb = xb.to(device)
+                xb = xb.to(device).float()
                 rates, kl_ic, kl_ctrl, _ = model(xb)
                 loss, _ = lfads_loss(rates, xb, kl_ic, kl_ctrl, kl_weight)
                 val_loss += loss.item()
